[FIX] 템플릿 메서드 패턴을 적용하여, stop() 메소드를 추상클래스에 구현
